requirements.txt
streamlit
python-dotenv
fdb
pandas